Initial implementation of upload_to_MOS
Initial implementation of the "upload_to_MOS" script, which uploads
diagnostic files to My Oracle Support (MOS) using the cURL utility...
this process is documented in Doc ID 1682567.1 on the MOS website. It's
primarily ueful for large files, which are not handled well (and often
fail to complete) via the MOS web interface.

Usage:

   upload_to_mos file1 [[file2]...] sr_number

The default MOS username is defined using the "mos_account=" line near
the top of the script, and can be overridden using the MOS_ACCOUNT
environment variable. You'll be prompted for your MOS password, which is
provided to cURL once for each file uploaded. This value isn't saved, or
remembered beyond the single script invocation.