Initial implementation of tnsping analogue
Initial implementation of the "dbping" script, which attempts to
approximate output of the Oracle "tnsping" utility. It has one major
difference in that it actually attempts to authenticate with the target
database using "sqlplus", which is expected to fail, whereas tnsping
never goes beyond the listener. So the times are expected to be a bit
longer than what tnsping would report, but the test is somewhat more
thorough as well.

Usage:

   dbping connection_string [count]

The first parameter is required, and can be any connection string which
sqlplus is configured to recognize... typically a tnsnames entry or
EZConnect string. The second is an optional count, defaulting to 1,
which indicates how many times the test should be performed.