Update execution plan display format
Update the display format passed by default to DBMS_XPLAN.DISPLAY, to
make it more useful for the common case. The output will now include
runtime rowcount info when available, which is essential for identifying
cardinality mismatch issues. This requires the SQL statement to either
include the GATHER_PLAN_STATISTICS hint, or to be executed with the
STATISTICS_LEVEL parameter set to ALL at either the session or system
level.