Add parallel server ID to blocker tree output
Add a PQ column, which indicates the parallel query server (P001, for
example) when applicable. This makes it easier to determine when one
thread of a parallel query is blocking another, which is part of normal
parallel operation... this is very likely to be the case when both
USERNAME and SQL_ID match within a given branch. Most people wouldn't
consider this a *true* blocking lock, since it isn't affecting other
sessions.

Example output (with column lengths shortened):

SQL> @blocker_tree

Lock Tree                 USERNAME        SQL_ID        STATUS   PQ
------------------------- --------------- ------------- -------- ----
1.2448.49655              SYSTEM          1crqrd757d0mf ACTIVE   P002
  1.4313.15755            SYSTEM          1crqrd757d0mf ACTIVE