Add scripts to convert hex value to DATE/TIMESTAMP
Add a series of "hex_to_*.sql" scripts, which convert a hexadecimal
number to its DATE or TIMESTAMP equivalent and displays the results
using the session NLS settings. This is useful for interpreting values
found in tracefiles and SQL Monitor.

Sample output:

   SQL> @hex_to_date 7877090d01363a

   DATE
   -------------------
   2019-09-13 00:53:57

   SQL> @hex_to_timestamp 7877090d01363a008efd78

   TIMESTAMP
   --------------------------
   2019-09-13 00:53:57.009371

   SQL> @hex_to_timestamp_tz 7877090d05363a008efd78103c

   TIMESTAMP_TZ
   ---------------------------------
   2019-09-13 00:53:57.009371 -04:00

The following article provided a great deal of helpful information
concerning the internal DATE/TIMESTAMP storage format.

   https://oracle-base.com/articles/misc/oracle-dates-timestamps-and-intervals