Fix detection of parallel query slaves
Oracle parallel query slave processes are normally named using the
pattern Pnnn, which needs to be parsed from the program name. Until now
we've been incorrectly expecting nnn to be a decimal number, when it's
actually implemented as a hexadecimal value. An additional wrinkle is
that query slaves spawned for GV$ views will instead be named like PZnn
or PPnn, depending upon the database version.